pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mu-dsu"
version = "0.1.0"
description = "μ-DSU: Micro-language based Dynamic Software Updating framework. Reimplementation of Cazzola et al. (2017)."
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "lark>=1.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0",
]

[project.scripts]
mu-dsu = "mu_dsu.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "mypy>=1.10",
    "ruff",
]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.ruff]
target-version = "py312"
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
